+++ v2/yarpcfx/yarpcfxmiddleware/inbound_encoding_test.go
@@ -18,7 +18,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package yarpcmiddlewarefx
+package yarpcfxmiddleware
 
 import (
 	"strings"
@@ -30,51 +30,51 @@ import (
 	yarpc "go.uber.org/yarpc/v2"
 )
 
-func TestNewInboundTransportConfig(t *testing.T) {
-	cfg := strings.NewReader(`yarpc: {middleware: {inbounds: {transport: {unary: ["nop"]}}}}`)
+func TestNewInboundEncodingConfig(t *testing.T) {
+	cfg := strings.NewReader(`yarpc: {middleware: {inbounds: {encoding: {unary: ["nop"]}}}}`)
 	provider, err := config.NewYAML(config.Source(cfg))
 	require.NoError(t, err)
 
-	res, err := NewInboundTransportConfig(InboundTransportConfigParams{
+	res, err := NewInboundEncodingConfig(InboundEncodingConfigParams{
 		Provider: provider,
 	})
 	require.NoError(t, err)
-	assert.Equal(t, InboundTransportConfig{Unary: []string{"nop"}}, res.Config)
+	assert.Equal(t, InboundEncodingConfig{Unary: []string{"nop"}}, res.Config)
 }
 
-func TestNewUnaryInboundTransport(t *testing.T) {
+func TestNewUnaryInboundEncoding(t *testing.T) {
 	t.Run("duplicate registration error", func(t *testing.T) {
-		_, err := NewUnaryInboundTransport(
-			UnaryInboundTransportParams{
-				Middleware: []yarpc.UnaryInboundTransportMiddleware{
-					yarpc.NopUnaryInboundTransportMiddleware,
-					yarpc.NopUnaryInboundTransportMiddleware,
+		_, err := NewUnaryInboundEncoding(
+			UnaryInboundEncodingParams{
+				Middleware: []yarpc.UnaryInboundEncodingMiddleware{
+					yarpc.NopUnaryInboundEncodingMiddleware,
+					yarpc.NopUnaryInboundEncodingMiddleware,
 				},
 			},
 		)
-		assert.EqualError(t, err, `unary inbound transport middleware "nop" was registered more than once`)
+		assert.EqualError(t, err, `unary inbound encoding middleware "nop" was registered more than once`)
 	})
 
 	t.Run("configured middleware is not available", func(t *testing.T) {
-		_, err := NewUnaryInboundTransport(
-			UnaryInboundTransportParams{
-				Config: InboundTransportConfig{
+		_, err := NewUnaryInboundEncoding(
+			UnaryInboundEncodingParams{
+				Config: InboundEncodingConfig{
 					Unary: []string{"dne"},
 				},
 			},
 		)
-		assert.EqualError(t, err, `failed to resolve unary inbound transport middleware: "dne"`)
+		assert.EqualError(t, err, `failed to resolve unary inbound encoding middleware: "dne"`)
 	})
 
 	t.Run("successful construction", func(t *testing.T) {
-		res, err := NewUnaryInboundTransport(
-			UnaryInboundTransportParams{
-				Config: InboundTransportConfig{
+		res, err := NewUnaryInboundEncoding(
+			UnaryInboundEncodingParams{
+				Config: InboundEncodingConfig{
 					Unary: []string{"nop"},
 				},
-				MiddlewareLists: [][]yarpc.UnaryInboundTransportMiddleware{
+				MiddlewareLists: [][]yarpc.UnaryInboundEncodingMiddleware{
 					{
-						yarpc.NopUnaryInboundTransportMiddleware,
+						yarpc.NopUnaryInboundEncodingMiddleware,
 					},
 				},
 			},
