+++ v2/yarpcfx/yarpcfxmiddleware/module.go
@@ -18,10 +18,21 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package yarpcmiddlewarefx
+package yarpcfxmiddleware
 
-const (
-	outboundTransportConfigurationKey = "yarpc.middleware.outbounds.transport"
-	inboundTransportConfigurationKey  = "yarpc.middleware.inbounds.transport"
-	inboundEncodingConfigurationKey   = "yarpc.middleware.inbounds.encoding"
+import (
+	"go.uber.org/fx"
+)
+
+// Module produces ordered slices of middleware according to
+// the middleware configuration.
+var Module = fx.Provide(
+	NewOutboundTransportConfig,
+	NewUnaryOutboundTransport,
+
+	NewInboundTransportConfig,
+	NewUnaryInboundTransport,
+
+	NewInboundEncodingConfig,
+	NewUnaryInboundEncoding,
 )
